requirements.txt
streamlit
python-dotenv
langchain==0.1.16
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langchain-google-genai==1.0.4
google-generativeai==0.5.4
sentence-transformers
huggingface-hub
torch
faiss-cpu
pypdf